requirements.txt
django==3.2
djangorestframework==3.13.1
drf-yasg==1.20.0
pytest==7.1.1
requests~=2.27.1
